checkin of classroom updates.
