+++ improvement_analysis/github-classroom-drivetrain-talon/template/src/main/java/frc/robot/subsystems/CANDriveSubsystem.java
@@ -6,8 +6,8 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
-// TODO Unit 1: Add imports for TalonFX, DifferentialDrive using VSCode quick-fix (Ctrl+.)
-// TODO Unit 2: Add import for TalonFXConfiguration
+// TODO Unit 1: Add import for TalonFX using VSCode quick-fix (Ctrl+.)
+// TODO Unit 2: Add imports for TalonFXConfiguration and DifferentialDrive
 // TODO Unit 2: Add static import for DriveConstants: import static frc.robot.Constants.DriveConstants.*;
 // TODO Unit 3: Add imports for Follower, InvertedValue, Command, and DoubleSupplier
 
@@ -16,26 +16,14 @@ public class CANDriveSubsystem extends SubsystemBase {
   // TODO Unit 1: Declare four private final TalonFX fields:
   //   leftLeader, leftFollower, rightLeader, rightFollower
 
-  // TODO Unit 1: Declare a private final DifferentialDrive field named drive
+  // TODO Unit 2: Declare a private final DifferentialDrive field named drive (above the constructor)
 
   /** Creates a new CANDriveSubsystem. */
   public CANDriveSubsystem() {
-    // TODO Unit 2: Initialize the four TalonFX objects using DriveConstants CAN IDs
-    //              Example: leftLeader = new TalonFX(LEFT_LEADER_ID);
-
-    // TODO Unit 2: Create a TalonFXConfiguration and set current limits on all motors
-    //              Example: config.CurrentLimits.SupplyCurrentLimit = DRIVE_MOTOR_CURRENT_LIMIT;
-    //              Then apply with getConfigurator().apply(config);
-
-    // TODO Unit 2: Initialize the drive field with DifferentialDrive(leftLeader, rightLeader)
-    //              Place this right after motor creation, before configuration
-
-    // TODO Unit 3: Set up rightFollower to follow rightLeader using Follower control class
-
-    // TODO Unit 3: Set up leftFollower to follow leftLeader using Follower control class
-
-    // TODO Unit 3: Configure left side inversion using InvertedValue.Clockwise_Positive
-    //              Apply the config to both leftLeader and leftFollower
+    // Placeholder — replace this entire constructor body in Units 2 and 3.
+    // This throw allows the class to compile while the final fields are unassigned.
+    // See unit2-constructor-constants.md and unit3-arcade-drive-factory.md for instructions.
+    throw new UnsupportedOperationException("CANDriveSubsystem not yet implemented — complete Units 2 and 3");
   }
 
   // TODO Unit 3: Add the driveArcade command factory method below.
