+++ improvement_analysis/github-classroom-drivetrain-talon/template/unit1-declare-fields.md
@@ -1,6 +1,6 @@
 # Unit 1: Declare Motor Variables
 
-**Objective:** Declare four `TalonFX` motor fields and a `DifferentialDrive` instance in the `CANDriveSubsystem` class.
+**Objective:** Declare four `TalonFX` motor fields in the `CANDriveSubsystem` class. The `DifferentialDrive` field will be added in Unit 2, alongside its initialization.
 
 **Estimated time:** 5 minutes
 
@@ -10,7 +10,7 @@ Open `src/main/java/frc/robot/subsystems/CANDriveSubsystem.java` and add the mot
 
 ## Instructions
 
-1. Add imports for `TalonFX` (from `com.ctre.phoenix6.hardware`) and `DifferentialDrive` (from `edu.wpi.first.wpilibj.drive`) at the top of the file. VSCode can add these automatically — hover over the red-underlined class name and choose **Quick Fix → Import**.
+1. Add an import for `TalonFX` (from `com.ctre.phoenix6.hardware`) at the top of the file. VSCode can add this automatically — hover over the red-underlined class name and choose **Quick Fix → Import**.
 
 2. Declare four `private final TalonFX` fields for the motors:
 
@@ -21,7 +21,7 @@ Open `src/main/java/frc/robot/subsystems/CANDriveSubsystem.java` and add the mot
 
    Each field is declared with its type and name only — do not assign a value here. The value is set in the constructor (Unit 2).
 
-3. Declare a `private final DifferentialDrive` field named `drive`. Like the motor fields, leave it unassigned for now.
+> **Note:** The `DifferentialDrive drive` field will be declared and initialized together in Unit 2, since it needs the motor objects to exist first.
 
 See the **Declaring Motor Variables** section of the [Driving Robot tutorial](../../docs/programming/driving_robot.md) for the TalonFX field declaration pattern, and the [full TalonFX subsystem example](../../docs/code_examples/2026KitBotInlineTalonFX/subsystems/CANDriveSubsystem.java) to check your work.
 
@@ -29,7 +29,6 @@ See the **Declaring Motor Variables** section of the [Driving Robot tutorial](..
 
 - **`private final`** — prevents accidental reassignment and limits scope to this subsystem
 - **TalonFX fields** — represent the physical motors; you'll initialize them in Unit 2
-- **DifferentialDrive** — provides arcade-drive logic that converts joystick input into left/right motor speeds
 
 ## Key Concepts
 
@@ -38,6 +37,8 @@ See the **Declaring Motor Variables** section of the [Driving Robot tutorial](..
 ## Verification
 
 After adding the declarations, your code should compile without errors. In VSCode, confirm there are no red underlines on the new field declarations. If you see an error about an unresolved symbol, use the Quick Fix import option or verify that Phoenix 6 is installed via the WPILib Command Palette → **Manage Vendor Libraries**.
+
+> **Tip:** The constructor currently has a `throw` placeholder that lets the class compile even though the `final` fields aren't assigned yet. You will replace the entire constructor body in Unit 2.
 
 ## Next Step
